Extract AvailableSlots to core/services
Pairing session with Tech Lead
